+++ ios/Plugin/Extensions/CapacitorVideoPlayerPlugin.Fullscreen.swift
@@ -19,8 +19,11 @@ extension CapacitorVideoPlayerPlugin {
     func createVideoPlayerFullscreenView(
         call: CAPPluginCall, videoUrl: URL, rate: Float,
         exitOnEnd: Bool, loopOnEnd: Bool, pipEnabled: Bool,
-        backModeEnabled: Bool, subTitleUrl: URL?,
-        subTitleLanguage: String?, subTitleOptions: [String: Any]?, headers: [String: String]?) {
+        backModeEnabled: Bool, showControls: Bool,
+        displayMode: String,
+        subTitleUrl: URL?, subTitleLanguage: String?,
+        subTitleOptions: [String: Any]?,
+        headers: [String: String]?) {
         DispatchQueue.main.async { [weak self] in
             let playerId: String = self?.fsPlayerId ?? "fullscreen"
             if let fullscreenView = self?.implementation
@@ -28,6 +31,8 @@ extension CapacitorVideoPlayerPlugin {
                     playerId: playerId, videoUrl: videoUrl,
                     rate: rate, exitOnEnd: exitOnEnd, loopOnEnd: loopOnEnd,
                     pipEnabled: pipEnabled,
+                    showControls: showControls,
+                    displayMode: displayMode,
                     subTitleUrl: subTitleUrl,
                     language: subTitleLanguage, headers: headers, options: subTitleOptions) {
                 self?.videoPlayerFullScreenView = fullscreenView
@@ -44,39 +49,39 @@ extension CapacitorVideoPlayerPlugin {
                     return
                 }
                 videoPlayer.delegate = self
-
-                self?.bridge?.viewController?.present(videoPlayer, animated: true, completion: {
-                    if backModeEnabled {
-                        // add audio session
-                        self?.audioSession = AVAudioSession.sharedInstance()
-                        // Set the audio session category, mode, and options.
-                        do {
-                            try self?.audioSession?
-                                .setCategory(.playback,
-                                             mode: .default,
-                                             options: [.mixWithOthers])
-                            try self?.audioSession?.setActive(true)
+                self?.bridge?.viewController?.present(
+                    videoPlayer, animated: true, completion: {
+                        if backModeEnabled {
+                            // add audio session
+                            self?.audioSession = AVAudioSession.sharedInstance()
+                            // Set the audio session category, mode, and options.
+                            do {
+                                try self?.audioSession?
+                                    .setCategory(.playback,
+                                                 mode: .default,
+                                                 options: [.mixWithOthers])
+                                try self?.audioSession?.setActive(true)
+                                call.resolve([
+                                                "result": true,
+                                                "method": "createVideoPlayerFullScreenView",
+                                                "value": true])
+                                return
+                            } catch let error as NSError {
+                                print("Unable to activate audio session:  \(error.localizedDescription)")
+                                call.resolve([
+                                                "result": false,
+                                                "method": "createVideoPlayerFullScreenView",
+                                                "message": error.localizedDescription])
+                                return
+                            }
+                        } else {
                             call.resolve([
                                             "result": true,
                                             "method": "createVideoPlayerFullScreenView",
                                             "value": true])
                             return
-                        } catch let error as NSError {
-                            print("Unable to activate audio session:  \(error.localizedDescription)")
-                            call.resolve([
-                                            "result": false,
-                                            "method": "createVideoPlayerFullScreenView",
-                                            "message": error.localizedDescription])
-                            return
                         }
-                    } else {
-                        call.resolve([
-                                        "result": true,
-                                        "method": "createVideoPlayerFullScreenView",
-                                        "value": true])
-                        return
-                    }
-                })
+                    })
 
             }
         }