+++ ios/Plugin/VideoPlayer/Extensions/PortraitAVPlayerController.swift
@@ -1,0 +1,12 @@
+//
+//  PortraitAVPlayerController.swift
+//  CapacitorVideoPlayer
+//
+//  Created by  Quéau Jean Pierre on 23/05/2023.
+//
+
+import AVKit
+class PortraitAVPlayerController: AVPlayerViewController {
+    override var supportedInterfaceOrientations: UIInterfaceOrientationMask {
+        return .portrait    }
+}